requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langgraph==0.1.1
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
tavily-python==0.3.3
redis[asyncio]==5.0.4
structlog==24.1.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dotenv==1.0.1